FastAPI/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-multipart
python-dotenv
sqlalchemy